requirements.txt
Django==3.2.8
django-environ==0.7.0
djangorestframework==3.12.4
psycopg2-binary==2.9.1
drf-yasg[validation]==1.20.0
